rename parameter_mapping -> mapping
